nit code refactor
